+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0]
+
+### Added
+- Rate limiting
+
+### Changes
+- Updated unit tests for controllers to ignore rate limiting
+
 ## [0.2.0]
 
 ### Added